fix(portal): fix key error when creating gateway token (#11420)
When creating a gateway token we were attempting to set the key
`secret_nonce`, however, gateway tokens don't have a nonce value. This
attempt to set the nonce key resulted in a crash, but only after the
token had already been created in the DB. When the process crashed, the
liveview reloaded the page and started the process all over again.