fix(connlib): allow connecting to multiple Gateways in a site (#11397)
Further testing and thinking has uncovered that the approach from #11267
is fundamentally flawed. The list of Gateways that we send to the portal
is merely a preference. There are multiple reasons why the portal may
not honor this request and just send us a different Gateway. The most
obvious one is that the Gateway is in fact no longer online!

Therefore, the correct thing to do here is to just connect to the new
Gateway that we are given, even if it means that we are now connected to
two in the same site (after all, there may just be a network-partition
between Gateway and portal but not Gateway and Client). This does not
affect other resources because we track the Gateway to use on a
per-Resource level.

With this insight, we can actually clean-up how we compute the list of
preferred Gateways:

1. We used to sort this list lexically by the use of a `BTreeSet`. This
makes absolutely no sense.
2. We used to always send the same list, irrespective of which resource
we are trying to connect to.
a. Given that we track Gateways on a per-Resource basis, it makes sense
to vary the ordering of this list with the resource.
b. We now prefer the Gateway that we have previously been authorized
for.
c. In case this resource has never been authorized before, we prefer
Gateways that we are already connected to.
3. We use to keep a separate cache for the recently connected Gateways.
This is now gone.

In order to enable the above state tracking, we now _no longer_ clear
the `authorized_resources` list upon `reset`. This was unnecessary and
with the new logic, would actually have prevented from "sticking" to the
same Gateway upon roaming. The assumption here is that when users roam,
they are very likely still in a location that is physically close to the
previous one, i.e. going from WiFi to Data or vice versa. Thus,
connected to the same Gateway should still give you a reasonably good
connection.

One of the magical things of Firezone is that we can in fact roam TCP
connections IF we re-connect to the same Gateway because the layer-3
connection through Firezone stays intact, even if underlying network
connection changes. To allow for that, we need to make sure that we at
least attempt to re-connect to the same Gateway.