fix(connlib): allow concurrenct `create_flow` messages (#11438)
This effectively reverts #11266 minus the refactorings that were
introduced as part of that. The reason is already explained in #11407
but in short, I think that this was actually a bad design because it
relies on retransmissions of packets to establish new connections so the
time-to-first-byte increases for all but the "first" flow if we want to
concurrently connect to resources in the same site.

We now do allow concurrent connections to multiple Gateways as of
#11397, so it should be fine to race these messages.

Resolves: #11407