fix(connlib): don't queue up exact dups of control messages (#11440)
When we are network-partitioned from the portal, we are queuing up
control messages and send them as soon as we are back online. In case
the user tried to connect to something in the meantime, this meant that
likely identical `create_flow` messages got queued and are flushed to
the portal all at once. This is unnecessary, we only really need to send
one of those.

This semantic actually applies to all messages that we send. If we have
an exact duplicate of a message already queued, we do not need to queue
another one.