ci: include hash of all Rust source files in cache key (#11457)
Typically, only the hash of lockfiles is incorporated into the cache
key. In order to also cache the result of workspace crates to the point
where we do not rebuild anything at all, we need to incorporate the Rust
source files as well into the cache key. Otherwise, no new cache is
saved for changes to Rust files that don't touch the lockfiles.