fix(portal): delete client_tokens when account is disabled (#11469)
When an account is disabled, we need to:

- Disconnect all clients
- Show a banner
- Continue to allow admins to sign in and Gateways to remain connected

This PR fixes an issue where clients weren't disconnected, and adds
tests to ensure all of the above.

We also apply a minor refactor in this PR to ensure portal_sessions can
only be created by `account_admin_user` actor types.

Fixes #11108