ci: call actionlint from pre-commit (#11439)
This way we can run actionlint locally.

Additionally, we can can reduce the number of jobs and pin the
`actionlint` version more easily (the previous CI pin was pinning only
the github workflow, not the actionlint itself, which was always
'latest').

Pyflakes dependency is for actionlint linting embedded Python code
(optional dependency of actionlint).