fix(portal): always use 'true' for checkbox value (#11471)
When the live_table component was introduced, it exposed a `value`
attribute to be passed into a checkbox in order to use for a
multi-select style live table filter.

But those filters never got used.

What also happened is it introduced a bug wherein if you don't
explicitly pass a `value="true"`, the checkbox state would be lost on
LiveView DOM updates since _another_ change we made to all `input`
components just above would set `value` to `""` if it was `nil` or not
provided.

This meant that all checkboxes that don't explicitly pass `value="true"`
would effectively have a bug where they lose their checked state on any
form change.

Since we don't have a multi-select type of live table filter (and let's
be honest, we need to rip it out anyway), we can unwind this ball of
yarn by deleting the relevant live_table code and reverting checkboxes
to always send `value="true"`.

Fixes #11418

---------

Signed-off-by: Jamil <jamilbk@users.noreply.github.com>