fix(portal): use Req to fetch component versions (#11485)
Due to the nature of Finch pools and how we start them with
`rest_for_one` instead of `one_for_one`, whenever we deploy, the
`:shutdown` message needs to be explicitly handled, otherwise we get an
exception.

To fix this, we can ditch `Finch` entirely for the ComponentVersions
GenServer and use our shiny new `Req` dependency, which has its own
internal request pool we don't need to manage.