fix(portal): add missing index for groups.name (#11497)
Now that members are not maintained for the everyone group, determining
which resources to load for its members needs to hit an index for the
group.

To fix this, we index groups on `account_id, name, type` and then ensure
that all of the joins involved use `account_id` since all pkey indexes
have this as the first element in the composite index.

This query is critical as it's used in the hot path when setting up a
new resource connection whenever the cache is not hit.

Related: #11493