test(portal): implement oidc_controller tests (#11501)
Due to the critical nature of authentication, we implement a robust
suite of controller tests for the OIDC flow. The goal in this PR was to
reach 100% line coverage.

Notably, we ensure that the claims / tokens returned from the IdP are
sane:

- all fields are validated for maximum length
- all fields are validated to be strings
- we show generic errors and log more detailed information to aid in
troubleshooting for things like a missing `email` claim

All error conditions should now never crash the controller, preventing
the user from seeing any 500 errors, and logging an info log instead.


Related: #11056