fix(portal): don't query DB to find online relays (#11504)
The nice thing about Phoenix Presence is that we get a cluster-wide CRDT
that remains intact during deploys. We can save all of the metadata we
need for each of the relays connected so that clients and gateways don't
need to query the DB to find connected Relays.

These queries were number 2 on our highest-load queries list according
to CloudSQL insights. Whenever relays reconnected, thousands of queries
would hit this table to resolve online relays due to the way we react to
relay joins and leaves in the channel modules.