fix(portal): link channel pid to transport pid (#11507)
When a channel pid crashes, the transport pid (responsible for the
WebSocket) does not automatically die with it. This is to prevent other
multiplexed channels from being taken down.

Since we do not use other channels, this creates a problem because
connlib happily stays connected to a WebSocket with no Phoenix channel
alive in it.

Since heartbeats are managed in the transport pid (otherwise the number
of heartbeats would grow with the number of multiplexed channels), this
means the connection is effectively a zombie.

To fix this, we simply link the channel pid with the transport pid using
[`Process.link/1`](https://hexdocs.pm/elixir/Process.html#link/1), which
tears the transport down with the channel whenever it exits.

Fixes #11495