refactor(portal): use dedicated /readyz probe (#11533)
We've been having thundering herd issues with each deploy. This is due
to a few different reasons:

1. We use the same health check for instance liveness and application
readiness. The first one is used by the instance group manager to
determine when the instance / VM is "alive", and the second used by the
load balancer to determine which nodes to route traffic to.
Unfortunately, when deploying, this endpoint stays green for the load
balancer even for old nodes, which means flapping as connections hop
between old nodes and new.
2. All nodes are spun up, and then all nodes destroyed at once. This
causes a swarm of reconnects as all nodes are taken down roughly at the
same time.

To fix the above issues we:

1. Introduce a separate `/readyz` and `/healthz` readiness check. When
we deploy, the MIG will execute our shutdown script which touches
`/var/run/firezone-draining`. In the portal, we then return `503` on
`/readyz` if this file exists, causing the load balancer to stop routing
new traffic to this node. At 30s, the LB will cut all websockets to this
node, and at 40s the service will receive a SIGTERM to gracefully shut
down the elixir app.
2. We update the `update_policy` to roll over one node at a time instead
of 4 up and then 4 down. This smooths over the reconnects so as not to
swarm all at once.