fix(portal): don't disconnect pending relays on join (#11540)
When a Relay temporarily disconnects (such as during deploy), we add it
to a pending_leaves buffer so that if it rejoins with the same
stamp_secret (indicating the Relay _has_not_ restarted) we can "cancel"
the leave.

There was a bug where any relays in this list would be immediately
disconnected if an unrelated relay joined within the window. The chance
of this happening was not high, but also not negligible.

To prevent spurious `relays_presence` messages being pushed to the
clients/gateways, we cross-check that the relay id is in `joined_ids`,
that is, has been actually joined, before sending a disconnect for it.