fix(portal): remove relays_presence debouncer (#11543)
The relays_presence debouncer was implemented at a time before we had
the eBPF implementation and suspected control plane drops to be mostly
because of packet load.

After rolling out the eBPF codepath for all relaying scenarios and
verifying we are _not_ overloaded on the relay fleet, we are still
seeing large numbers of reconnects.

After testing locally and investigating logs in staging / production,
the leading hypotheses for the remaining sources of flaps are:

- Noisy join / leave events that may arrive out of order. These are
broadcasted by Phoenix.Presence as soon as an entity connects and
disconnects.
- A possible bug in the Relay eventloop that prevents a heartbeat from
being sent within the 30s Phoenix Channel heartbeat timeout, causing the
portal to disconnect this relay.

For scenario (1) above, we should be using `Presence.list/1` to query
the state of the CRDT instead of trying to maintain our own state from
received events. We can simply use the received event to trigger a
lookup in this managed state which is already debounced / merged /
reodered.

For scenario (2) above, further investigation is underway, but #11542 is
a candidate fix for the suspected issue there.