chore(portal): warn before relay channel idle timeouts (#11546)
Unfortunately we don't get any meaningful info when our channel pid is
cut due to transport timeout. However, we can poll `cowboy`'s state,
which uses a tick-based system, to detect when are about to be closed
due to missing a heartbeat.

This will allow us to better understand whether the WebSocket connection
cuts seen by the Relay are due to missing heartbeats or some other
cause.

Related: #11542