fix(portal): delay reactions to relay presence events (#11549)
Unfortunately WebSocket connections aren't always reliable. It seems GCP
or another thing in the path between relay and load balancer is
occasionally dropping connections, and this is not due to the 86400
`timeout_sec` maximum connection time in the GCP ALB, or (most likely)
not a heartbeat timing issue in the Relay.

Before, we had a complex state tracking mechanism which was tricky to
test and added much complexity to the already complex state tracking in
the channel modules. This was removed in #11543 in favor of looking the
"state of truth" from the Presence CRDT.

This new lookup is being extended in this PR to implement a much simpler
debouncer system.

This works by delaying all reactions to Presence joins/leaves by 1s so
that if a Relay disconnects and then reconnects, by the time we check
its Presence it will appear to never have disconnected.

Supersedes: #11542