refactor(portal): check billing limits more efficiently (#11550)
We need to regularly process accounts to limits based on what plan they
are on and then set a warning banner on the account if it's over.

Before, we'd loop over each account individually and run the aggregation
function. While not an immediate issue, this did cause a CPU spike every
30m when the job ran.

We can be _much_ more efficient here by letting the DB do the heavy
lifting here, performing this query as a batch for all accounts.