fix(portal): delete sessions for external_identities (#11564)
When an `external_identity` is deleted, it's a good idea to immediately
clean up any associated `portal_session`s and `client_token`s as well.

In practice, most of the deletions here will happen during sync, in
which the actor will be cleaned up as well triggering the cascade
delete. But this is good insurance.

Fixes #11408