fix(portal): prevent two identical relays from connecting (#11567)
Load balancers are interesting creatures. What appears to be happening
with ours is that the public-facing, downstream connection is severed
randomly but the backend, upstream connection remains intact.

This causes the Relay channel to disconnect eventually due to heartbeat,
but the Relay itself has actually already reconnected with a new healthy
connection.

The effect here is that we have _two_ identical Relays in the Presence
state, one healthy, and one a zombie that no one knows is a zombie, so
newly-joining clients/gateways might accidentally pick it to use for
connections.

To fix this, we immediately disconnect all other channels with the
stamp_secret of the Relay that's joining, to ensure only the most
recently connected Relay is used.