refactor(portal): ephemeral relays (#11587)
There is no reason we need persistence for Relays, which come and go and
are already tracked in the Presence state. Each client and gateway
subscribes to a debounced stream of events for updates to this state so
that they have up-to-date information on the list of relays available
for them to use.

The one thing we lose here is the database ID, a UUIDv4. Since any ID
will do here, this is now computed deterministically from the
stamp_secret, so that clients and gateways can differentiate relays with
different in-memory states (and therefore credentials) from each other.